Initial upload for examine_epidermal_thickness
An initial upload for a script which moves through the image data and
calculates the thickness of each tissue layer, and the full epidermis,
and then plots the results.

This script is modified from the version which produced Figure S6 from
Cursons et al (2015), BMC Sys. Bio., and it is used to produce Fig.
AF1.3 from Cursons et al (2015), GigaScience.